Create common class
