+++ tests/test_pr.py
@@ -3,28 +3,7 @@
 from gidgethub import sansio
 
 from webservice import pr, consts
-
-
-class FakeGH:
-    """Create fake GitHub object"""
-
-    def __init__(self):
-        self.post_data = None
-        self.delete_url = None
-        self.post_url = None
-        self.patch_url = None
-
-    async def delete(self, url):
-        """Trigger delete method"""
-        self.delete_url = url
-
-    async def post(self, url):
-        """Trigger post method"""
-        self.post_url = url
-
-    async def patch(self, url):
-        """Trigger patch method"""
-        self.patch_url = url
+from .gh import FakeGH
 
 
 async def test_pr_opened():
